Moved common compose functions to core ui
