--- feature/player/src/main/java/com/arcticoss/nextplayer/feature/player/composables/AddLifecycleEventObserver.kt
@@ -1,23 +1,0 @@
-package com.arcticoss.nextplayer.feature.player.presentation.composables
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleEventObserver
-import androidx.lifecycle.LifecycleOwner
-
-@Composable
-fun AddLifecycleEventObserver(
-    lifecycleOwner: LifecycleOwner,
-    onLifecycleEvent: (Lifecycle.Event) -> Unit
-) {
-    DisposableEffect(lifecycleOwner) {
-        val observer = LifecycleEventObserver { _, event ->
-            onLifecycleEvent(event)
-        }
-        lifecycleOwner.lifecycle.addObserver(observer)
-        onDispose {
-            lifecycleOwner.lifecycle.removeObserver(observer)
-        }
-    }
-}
