+++ feature/player/src/main/java/com/arcticoss/nextplayer/feature/player/composables/PlayerUIMainControls.kt
@@ -1,4 +1,4 @@
-package com.arcticoss.nextplayer.feature.player.presentation.composables
+package com.arcticoss.nextplayer.feature.player.composables
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
